Merges r4580 from nscldaq-11.0-000-development branch into main for use in 11.0-rc19 deployment

+++ main/sbs/vmemodules/CModule32.h
@@ -79,7 +79,7 @@ protected:
   void Key(uint32_t offset) const; //!< write a key register.
 
   static uint32_t Offset(uint32_t off)  {
-    return off/sizeof(long);
+    return off/sizeof(uint32_t);
   }
   static void ThrowString(const char*  pLeader,
 		     const char*  pMessage) throw (std::string);
